\name{geeni-package}
\alias{geeni-package}
\alias{geeni}
\docType{package}

\title{
  Parallel implmentation of bounded memory GEE
}
\description{
  Illustrate concepts of deployable statistical methodology using GEE solution.
}
\details{
  See \code{packageDescription('geeni')} for package details. A
  useful starting point is the \code{\link{pargee}} manual page.
}
\author{
  VJ Carey <stvjc@channing.harvard.edu>

  Maintainer: Vince Carey <stvjc@channing.harvard.edu>
}
\seealso{
  Related in spirit to Lumley's biglm, bigglm, but statistics are incremented 
  by data groups rather than records.
}
\examples{
packageDescription('geeni')
}
\keyword{package}
